# [PyTorch] Minor optimizations to reduce CPU overheads in modules