--- a/transformer_engine/pytorch/module/base.py
+++ b/transformer_engine/pytorch/module/base.py
@@ -11,7 +11,7 @@ import socket
 import fcntl
 import struct
 from abc import ABC, abstractmethod
-from typing import Any, Callable, Dict, Generator, List, Optional, Tuple, Union
+from typing import Any, Callable, Dict, Generator, List, Optional, Set, Tuple, Union
 from contextlib import contextmanager
 
 import torch
@@ -408,13 +408,6 @@ class TransformerEngineBaseModule(torch.nn.Module, ABC):
         self._fp8_workspaces: Dict[str, Float8Tensor] = {}
         self.activation_dtype: Optional[torch.dtype] = None
 
-        # Fast variant of __setattr__
-        # Note: torch.nn.Module has custom __setattr__ that handles
-        # modules, parameters, and buffers. This is unnecessary
-        # overhead when setting plain attrs.
-        self._fast_setattr: Callable[Tuple[str, Any], None]
-        self._fast_setattr = super(torch.nn.Module, self).__setattr__
-
         # Fast getter for parameters
         # Note: torch.nn.Module does not store parameters like normal
         # attrs, but rather in a dict. When attempting to access, the
@@ -423,6 +416,27 @@ class TransformerEngineBaseModule(torch.nn.Module, ABC):
         # we know we are accessing a parameter.
         self._fast_get_param: Callable[str, torch.nn.Parameter]
         self._fast_get_param = self.__dict__["_parameters"].get
+
+    # Names of attributes that can be set quickly (see __setattr__
+    # method)
+    _fast_setattr_names: Set[str] = {
+        "activation_dtype",
+        "fp8",
+        "fp8_initialized",
+        "fp8_calibration",
+        "fp8_parameters",
+    }
+
+    def __setattr__(self, name: str, value: Any) -> None:
+        if name in TransformerEngineBaseModule._fast_setattr_names:
+            # torch.nn.Module has a custom __setattr__ that handles
+            # modules, parameters, and buffers. This is unnecessary
+            # overhead when setting plain attrs.
+            self.__dict__[name] = value
+        else:
+            # Default case
+            super().__setattr__(name, value)
+
 
     def adjust_amax_history_length(self, length: int, fwd: Optional[bool] = None) -> None:
         """Increase or decrease size of amax history based on given `length`.
@@ -606,7 +620,7 @@ class TransformerEngineBaseModule(torch.nn.Module, ABC):
         """Get activation data type for AMP."""
         # Native AMP (`torch.autocast`) gets highest priority
         if torch.is_autocast_enabled():
-            self._fast_setattr("activation_dtype", torch.get_autocast_gpu_dtype())
+            self.activation_dtype = torch.get_autocast_gpu_dtype()
             return
 
         # All checks after this have already been performed once, thus skip
@@ -620,7 +634,7 @@ class TransformerEngineBaseModule(torch.nn.Module, ABC):
                     "Data types for parameters must match when outside of autocasted region. "
                     f" Found input dtype: {dtype} and {name!r} dtype: {param.dtype}"
                 )
-        self._fast_setattr("activation_dtype", dtype)
+        self.activation_dtype = dtype
 
     def set_tensor_parallel_group(self, tp_group: Union[dist_group_type, None]) -> None:
         """
@@ -649,9 +663,9 @@ class TransformerEngineBaseModule(torch.nn.Module, ABC):
     # assume FP8 execution.
     def init_fp8_metadata(self, num_gemms: int = 1) -> None:
         """Initialize fp8 related metadata and tensors during fprop."""
-        self._fast_setattr("fp8_parameters", FP8GlobalStateManager.with_fp8_parameters())
-        self._fast_setattr("fp8", FP8GlobalStateManager.is_fp8_enabled())
-        self._fast_setattr("fp8_calibration", FP8GlobalStateManager.is_fp8_calibration())
+        self.fp8_parameters = FP8GlobalStateManager.with_fp8_parameters()
+        self.fp8 = FP8GlobalStateManager.is_fp8_enabled()
+        self.fp8_calibration = FP8GlobalStateManager.is_fp8_calibration()
         self.fp8_meta["fp8_checkpoint"] = self.fp8 or self.fp8_calibration
 
         if self.fp8_parameters and not self.fp8_initialized:
@@ -677,10 +691,10 @@ class TransformerEngineBaseModule(torch.nn.Module, ABC):
 
             # Allocate scales and amaxes
             self.init_fp8_meta_tensors()
-            self._fast_setattr("fp8_initialized", True)
+            self.fp8_initialized = True
         else:
             # If fp8 isn't enabled, turn off and return.
-            self._fast_setattr("fp8_initialized", False)
+            self.fp8_initialized = False
             return
 
     @contextmanager