… tests/pytorch/test_numerics.py
… transformer_engine/pytorch/module/rmsnorm.py
… transformer_engine/pytorch/transformer.py
… transformer_engine/pytorch/utils.py