# Description

We have observed that TE modules experience non-trivial CPU overhead, which often becomes a performance bottleneck in the forward pass of small models. For example, measuring the CPU runtime for Megatron-core modules with BF16 compute and TP=1:
- [`ColumnParallelLinear`](https://github.com/NVIDIA/Megatron-LM/blob/6b35ca80e8baca6f357b97304979e6b1c9a31899/megatron/core/tensor_parallel/layers.py#L649): 74 us per forward pass
- [`TEColumnParallelLinear`](https://github.com/NVIDIA/Megatron-LM/blob/6b35ca80e8baca6f357b97304979e6b1c9a31899/megatron/core/extensions/transformer_engine.py#L354): 140 us per forward pass

Unfortunately this overhead is distributed throughout the forward pass. Many basic PyTorch operations, e.g. getting attributes from `torch.Tensor`, involve O(1 us) overhead, so even basic checks to handle all of our advanced features will eventually add up to something non-trivial.

This PR makes a few minor optimizations:
- Avoid importing from `te.pytorch.cpu_offload` in every forward pass
- Memoize NCCL process group properties
- Avoid custom logic in `torch.nn.Module.__setattr__` when possible
- Avoid custom logic for accessing params in `torch.nn.Module` when possible
- Avoid accessing tensor attrs more than necessary

I see a 1.22x speedup, with 115 us per forward pass.

## Type of change

- [ ] Documentation change (change only to the documentation, either a fix or a new content)
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Infra/Build change
- [x] Code refractor

## Changes

- Avoid importing from `te.pytorch.cpu_offload` in every forward pass
- Memoize NCCL process group properties
- Avoid custom logic in `torch.nn.Module.__setattr__` when possible
- Avoid custom logic for accessing params in `torch.nn.Module` when possible
- Avoid accessing tensor attrs more than necessary

# Checklist:

- [x] I have read and followed the [contributing guidelines](https://github.com/NVIDIA/TransformerEngine/blob/main/CONTRIBUTING.rst)
- [x] The functionality is complete
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my feature works
- [ ] New and existing unit tests pass locally with my changes